requirements.txt
flask
langchain
openai
duckduckgo-search
python-dotenv
requests
